add optional coordinator pubkey to event
